Add abort condition to "date when you have X gems" calculator
If a user puts in unattainable values (i.e. Tier 1 all the things
and doesn't have any gems to start with, i.e. net loss) then this
calculation could spin on into infinity, causing browser freak out.
We now set an arbitrary limit of 5 years, if the desired amount of
gems can't be obtained within that time, we bail.